Remove progressbar
Couldn't load before the results came in so effectively useless.
