+++ main.py
@@ -10,56 +10,6 @@ from pyowm.commons.exceptions import NotFoundError as OWMNotFoundError
 from requests import Response
 from requests import get as requests_get
 
-
-class ProgressBar(Toplevel):
-    def __init__(self, parent: App, *args, **kwargs):
-        # Set up window
-        super().__init__(parent, *args, **kwargs)
-        self.title("Loading...")
-        self.resizable(False, False)
-
-        # Set up widgets
-        self.main_frame = Frame(self)
-        self.main_frame.grid()
-
-        main_label = Label(self.main_frame, text="Loading...", font="Helvetica 15 bold")
-        main_label.grid(padx=10, pady=10)
-
-        self.progressbar = Progressbar(
-            self.main_frame,
-            orient="horizontal",
-            length=200,
-            mode="indeterminate",
-            maximum=4,
-        )
-        self.progressbar.grid(padx=10, pady=10)
-
-        self.resize_app()
-
-    def resize_app(self) -> App:
-        """Use tkinter to detect the minimum size of the app, get the center of the screen, and place the app there."""
-
-        # TODO: Make a global function for this to remove boilerplate code
-        # Update widgets so minimum size is accurate
-        self.update_idletasks()
-
-        # Get minimum size
-        minimum_width: int = self.winfo_reqwidth()
-        minimum_height: int = self.winfo_reqheight()
-
-        # Get center of screen based on minimum size
-        x_coords = int(self.winfo_screenwidth() / 2 - minimum_width / 2)
-        y_coords = int(self.wm_maxsize()[1] / 2 - minimum_height / 2)
-
-        # Place app and make the minimum size the actual minimum size (non-infringable)
-        self.geometry(f"{minimum_width}x{minimum_height}+{x_coords}+{y_coords}")
-        self.wm_minsize(minimum_width, minimum_height)
-        return self
-
-    def set_progress(self, progress: int):
-        self.progressbar.step(progress)
-
-
 class App(Tk):
     def __init__(self):
         super().__init__()
@@ -147,7 +97,7 @@ class App(Tk):
         self.resize_app()
         self.deiconify()
 
-    def about(self) -> App:
+    def about(self) -> None:
         """Display a messagebox with information about the app."""
 
         messagebox.showinfo(
@@ -155,9 +105,8 @@ class App(Tk):
             "Weather is a simple weather app that uses the OpenWeatherMap API to get the weather for a given city.",
             parent=self,
         )
-        return self
 
-    def resize_app(self) -> App:
+    def resize_app(self) -> None:
         """Use tkinter to detect the minimum size of the app, get the center of the screen, and place the app there."""
 
         # Update widgets so minimum size is accurate
@@ -174,7 +123,6 @@ class App(Tk):
         # Place app and make the minimum size the actual minimum size (non-infringable)
         self.geometry(f"{minimum_width}x{minimum_height}+{x_coords}+{y_coords}")
         self.wm_minsize(minimum_width, minimum_height)
-        return self
 
     def exit_app(self) -> None:
         """Exit the app."""
@@ -192,7 +140,6 @@ class App(Tk):
         self.start_button.configure(state="disabled")
         self.searchbar.configure(state="disabled")
         self.update_labels()
-        pb = ProgressBar(self)
 
         # Get API key
         api_key: str = "c439e1209216cc7e7c73a3a8d1d12bfd"
@@ -209,12 +156,9 @@ class App(Tk):
             self.update_labels()
             self.start_button.configure(state="normal")
             self.searchbar.configure(state="normal")
-            pb.destroy()
             self.searching = False
             self.resize_app()  # In case the name gets too long or it renders differently on other systems
             return
-
-        pb.progressbar.step()
 
         # Check if city exists
         try:
@@ -224,12 +168,9 @@ class App(Tk):
             self.update_labels()
             self.start_button.configure(state="normal")
             self.searchbar.configure(state="normal")
-            pb.destroy()
             self.searching = False
             self.resize_app()  # In case the name gets too long or it renders differently on other systems
             return
-
-        pb.progressbar.step()
 
         # Get weather data
         weather = observation.weather
@@ -245,12 +186,9 @@ class App(Tk):
             self.update_labels()
             self.start_button.configure(state="normal")
             self.searchbar.configure(state="normal")
-            pb.destroy()
             self.searching = False
             self.resize_app()  # In case the name gets too long or it renders differently on other systems
             return
-
-        pb.progressbar.step()
 
         # Get response data, simplify and create variables for usage
         # TODO: Give all data in dual columns
@@ -278,14 +216,9 @@ class App(Tk):
         # Set the city name
         self.cityname.configure(text=f"City: {city}")
 
-        # Update the Progress Bar one last time
-        print("Updating Progress Bar one last time")
-        pb.progressbar.step()
-
         # Stop the Progress Bar, enable buttons and searchbar, and set searching to False
         self.start_button.configure(state="normal")
         self.searchbar.configure(state="normal")
-        pb.destroy()
         self.searching = False
         self.resize_app()  # In case the name gets too long or it renders differently on other systems
